% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_unscaled.R
\name{apply_universal_slope}
\alias{apply_universal_slope}
\title{Apply universal slope correction to DSI}
\usage{
apply_universal_slope(data)
}
\arguments{
\item{data}{A full DSI data set including the subset list-column}
}
\value{
The full DSI data set with the res_i_adj values replaced by their universal-slope equivalents and a new column, res_i_adj_s, with z-scored values.
}
\description{
Apply universal slope correction to DSI
}
