% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emmreml-related.R
\name{batch_emm}
\alias{batch_emm}
\title{batch_emm}
\usage{
batch_emm(
  info,
  variable,
  resids,
  rel,
  lps_name = "treatment",
  numcore = 3,
  sname = "sname.x",
  plot = F,
  savedir = "NA",
  x_axis_size = 20
)
}
\arguments{
\item{info}{meta info data frame containing lps and other variable}

\item{variable}{a vector of variables to be included in emmreml}

\item{lps_name}{name of "LPS"/"NULL" column, default to "treatment"}

\item{numcore}{default to 3}

\item{sname}{default to "sname.x"}

\item{plot}{plot p histogram or not, default to F}

\item{savedir}{if you want to save the picture, specify a dir}

\item{x_axis_size}{font size of the x axis title, default to 20}

\item{resids_female}{GE residuals data frame, will automatically be transformed to matrix}

\item{rel_female}{relatedness data frame, same as above}
}
\value{
a combined result datarame
}
\description{
run emmreml for multiple variable in one time
}
\examples{
batch_emm(meta_info,c("age","rank"),resids,rel)
}
