% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emmreml-related.R
\name{create_design}
\alias{create_design}
\title{create_design}
\usage{
create_design(data, nested_name, lps_name = "treatment")
}
\arguments{
\item{data}{dataframe containing meta info}

\item{nested_name}{var to be nested within LPS vs. Null}

\item{lps_name}{name of the column containing "LPS"/"NULL", default to "treatment"}
}
\value{
a design matrix
}
\description{
create_design
}
\examples{
create_design(dataframe,"rank","treatment")
}
