% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emmreml-pipeline.R
\name{emm_pip}
\alias{emm_pip}
\title{emm_pip}
\usage{
emm_pip(
  ge_meta,
  sci,
  agi,
  dsi,
  dsi_sum,
  type = "wt_avg",
  res,
  rel,
  sex = "F",
  fix_var = c("age", "rank"),
  change_var = c("eigen_wt", paste0(c("SCI", "AGI", "DSI", "SumBond"), "_F"),
    paste0(c("SCI", "AGI", "DSI", "SumBond"), "_M")),
  lps_name = "treatment",
  numcore = 3,
  sname = "sname",
  outputdir,
  emm_prefix,
  x_axis_size = 20
)
}
\arguments{
\item{ge_meta}{should contain sname,dart_date,treatment,group,sex,rank,age}

\item{sci}{sci input from Ramboseli::sci}

\item{agi}{agi input from Ramboseli::agi}

\item{dsi}{dsi_pop input from Ramboseli::dsi_pop}

\item{dsi_sum}{dsi_pop_sumamry input from Ramboseli::dsi_pop_summary}

\item{type}{"wt_avg"(default),"dir_avg", or "grp_only". calculating weighted average by days_present across group, direct average, or keep only the group observed on the darting date}

\item{res}{full GE residual dataframe}

\item{rel}{full K matrix}

\item{sex}{The sex to look at, "F" (default) or "M"}

\item{fix_var}{The variables to be included in every emmreml model, defaut to c("age","rank")}

\item{change_var}{The variables to be added with replacement to the basic model. Default to all SR.}

\item{lps_name}{The colname indicating LPS/NULL in ge_meta, default to treatment}

\item{numcore}{The number of core to be used for EMMREML, default to 3}

\item{sname}{The colname indicaiting sname, default to sname, could be sname.x}

\item{outputdir}{The folder to store output, end with /}

\item{emm_prefix}{The prefix for emmresult RDS and p histogram png}

\item{x_axis_size}{The font size of x axis of p histogram, default to 20}
}
\value{
a data frame with merged emmreml result

Nothing, results will be stored in the specified output directory
}
\description{
Pipeline for running emmreml
}
