% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visual-related.R
\name{gene_num_comparison}
\alias{gene_num_comparison}
\title{Title}
\usage{
gene_num_comparison(
  single_hallway,
  hallway_name,
  emm_list,
  var_list,
  suffix_list
)
}
\arguments{
\item{single_hallway}{a list element, a single hallway}

\item{hallway_name}{a string}

\item{emm_list}{a emmresult data frame list, list(1) or length=var_list}

\item{var_list}{a vector of variable name beta_var}

\item{suffix_list}{a vector of suffix}
}
\value{
a gg object
}
\description{
Title
}
