% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR-related.R
\name{get_centrality}
\alias{get_centrality}
\title{get_centrality}
\usage{
get_centrality(my_focal, my_grp, my_subset)
}
\arguments{
\item{my_focal}{my focal}

\item{my_grp}{my group}

\item{my_subset}{my subset}
}
\value{
eigenvector
}
\description{
get_centrality
}
