% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_unscaled.R
\name{get_focal_index}
\alias{get_focal_index}
\title{Calculate dyadic variables for focal animal from a dyadic subset}
\usage{
get_focal_index(my_sname, my_grp, my_subset)
}
\arguments{
\item{my_sname}{mysname}

\item{my_grp}{mygrp}

\item{my_subset}{mysubset}
}
\value{
The input data with an additional list column containing the dyadic variables.
}
\description{
Calculate dyadic variables for focal animal from a dyadic subset
}
