% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_unscaled.R
\name{get_interaction_dates}
\alias{get_interaction_dates}
\title{get_interaction_dates}
\usage{
get_interaction_dates(my_sub, members_l, df, my_sex_var, my_role, my_sex)
}
\arguments{
\item{my_sub}{god knows what}

\item{members_l}{god knows what}

\item{df}{god knows what}

\item{my_sex_var}{god knows what}

\item{my_role}{god knows what}

\item{my_sex}{god knows what}
}
\value{
god knows what
}
\description{
get_interaction_dates
}
