% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_unscaled.R
\name{get_mem_dates_sh}
\alias{get_mem_dates_sh}
\title{get_mem_dates_sh}
\usage{
get_mem_dates_sh(my_sub, members_l, df, sel = NULL)
}
\arguments{
\item{my_sub}{God knows what}

\item{members_l}{God knows what}

\item{df}{God knows what}

\item{sel}{God knows what}
}
\value{
God knows what
}
\description{
get_mem_dates_sh
}
