% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visual-related.R
\name{GSEA_dot_plot}
\alias{GSEA_dot_plot}
\title{Title}
\usage{
GSEA_dot_plot(gsea_list, var_vec, suffix_vec, pathway = c(1:50), labsize = 1)
}
\arguments{
\item{gsea_list}{a list of gsea results, 1 element or many}

\item{var_vec}{a vector of var, 1 element of many}

\item{suffix_vec}{the suffix, has to be multiple, will directly be x-axis-title}

\item{pathway}{the index of pathway to keep if only a subset}

\item{labsize}{the size of axis title, default to 1}
}
\value{
a gg object
}
\description{
Title
}
