% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA-pipeline.R
\name{gsea_pip}
\alias{gsea_pip}
\title{gsea_pip}
\usage{
gsea_pip(
  emm_result,
  varname = "all",
  hallway,
  numcore = 3,
  sim = 10000,
  type = NULL
)
}
\arguments{
\item{emm_result}{The emm result dataframe}

\item{varname}{A vector of varname, var_varname, beta_varname, default to all}

\item{hallway}{The list containing all hallway info}

\item{numcore}{Number of core, default to 3}

\item{sim}{number of simulation round}

\item{type}{add a column stating the info, default to NULL}
}
\value{
a data frame containing all ES_p and ES
}
\description{
gsea_pip
}
