% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA-related.R
\name{GSEA_var}
\alias{GSEA_var}
\title{GSEA_var}
\usage{
GSEA_var(emm_result, varname, hallway, numcore = 3, sim = 10000)
}
\arguments{
\item{emm_result}{emmreml result data frame}

\item{varname}{variable name for calculation, var_varname, beta_varname}

\item{hallway}{the list containing info about 50 hallways}

\item{numcore}{number of cores to use, default 3}

\item{sim}{number rounds of simulation, default 10000}
}
\value{
a dataframe containing ES and ES_p
}
\description{
GSEA_var
}
\examples{
GSEA_var(emm,"rank_lps")
}
