% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visual-related.R
\name{log10qq}
\alias{log10qq}
\title{log10qq}
\usage{
log10qq(
  data,
  var_list,
  colorlab = "variable",
  qlength = 1000,
  ptsize = 1,
  labsize = 5,
  alpha = 0.7
)
}
\arguments{
\item{data}{emm result frame}

\item{var_list}{a vector of varname, p_varname}

\item{colorlab}{label for color legend, default to variable}

\item{qlength}{resolution for quantile, default to 1000}

\item{ptsize}{point size, default to 1}

\item{labsize}{lab text size}

\item{alpha}{default to 0.7}
}
\value{
p a ggplot object
}
\description{
log10qq
}
