% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visual-related.R
\name{log10qq_acrossemm}
\alias{log10qq_acrossemm}
\title{Title}
\usage{
log10qq_acrossemm(
  emmlist,
  suffix_list,
  var_name,
  colorlab = "variable",
  qlength = 1000,
  ptsize = 1,
  labsize = 5,
  alpha = 0.7
)
}
\arguments{
\item{emmlist}{a list of emm result}

\item{suffix_list}{a list of suffix lps_null_suffix}

\item{var_name}{a pattern to grep "lps_null", an element or a list length = emmlist}

\item{colorlab}{label for color legend, default to variable}

\item{qlength}{resolution for quantile, default to 1000}

\item{ptsize}{point size, default to 1}

\item{labsize}{lab text size}

\item{alpha}{default to 0.7}
}
\value{
p a ggplot object
}
\description{
Title
}
