% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewrite_make_target_dates.R
\name{make_EL_target_date_df_sh}
\alias{make_EL_target_date_df_sh}
\title{Create a data frame with early life before age 5 or first start date of mother}
\usage{
make_EL_target_date_df_sh(
  target_df,
  babase,
  members_l,
  window_length = 1,
  window_length_day = 0,
  min_pres_day = 0,
  .by_grp = TRUE,
  .adults_only = TRUE,
  series = F,
  overlap = 0
)
}
\arguments{
\item{target_df}{A data frame that includes the columns sname, sex, grp, and date}

\item{babase}{A DBI connection to the babase database}

\item{members_l}{A subset of members table produced by the function 'subset_members'}

\item{window_length}{Length in years of the time window for the social index}

\item{window_length_day}{Length in days in addition to years of the time window}

\item{min_pres_day}{minimum number of days in a group, default to 0}

\item{.by_grp}{Logical indicating whether to separate by group. Default is TRUE}

\item{.adults_only}{Logical indicating whether to include adults only. Default is TRUE}

\item{series}{logical indicating whether to create a series instead of a single value}

\item{overlap}{in term of days, default to zero}
}
\value{
A tibble with one row per animal (and optionally, per group) and target date, with contextual data
}
\description{
Create a data frame with early life before age 5 or first start date of mother
}
