% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR-related.R
\name{make_target_date_custom_df}
\alias{make_target_date_custom_df}
\title{Create a data frame with year-long intervals given start and end date, and backwards.}
\usage{
make_target_date_custom_df(
  target_df,
  babase,
  members_l,
  window_shift = 0,
  window_length = 1,
  .by_grp = TRUE,
  .adults_only = TRUE,
  .early_life = FALSE,
  .cumulative_adult_life = F
)
}
\arguments{
\item{target_df}{A data frame that includes the columns sname, sex, grp, and date}

\item{babase}{A DBI connection to the babase database}

\item{members_l}{A subset of members table produced by the function 'subset_members'}

\item{window_shift}{Length in years of the time shift for the social index. i.e. 1: 2-1 years prior}

\item{window_length}{Length in years of the time window for the social index}

\item{.by_grp}{Logical indicating whether to separate by group. Default is TRUE}

\item{.adults_only}{Logical indicating whether to include adults only. Default is TRUE}

\item{.early_life}{Logical indicating whether to calculate early life: birth - rnk/mature date, default is FALSE}

\item{.cumulative_adult_life}{Logical indicating whether to calculate adult life: rnk/mature date - darting date, default is FALSE}
}
\value{
A tibble with one row per animal (and optionally, per group) and target date, with contextual data
}
\description{
Create a data frame with year-long intervals given start and end date, and backwards.
}
