% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR-related.R
\name{merge_sr}
\alias{merge_sr}
\title{merge_sr}
\usage{
merge_sr(sci, agi, dsi_pop, dsi_pop_summary)
}
\arguments{
\item{sci}{SCI data frame}

\item{agi}{agi data frame}

\item{dsi_pop}{dsi_pop data frame}

\item{dsi_pop_summary}{dsi_pop_summary data frame}
}
\value{
a merged social info data frame
}
\description{
Merging sci, agi, dsi_pop and dsi_pop_summary, calculate eigen_wt and sumbond
}
