% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emmreml-related.R
\name{perm.fdr}
\alias{perm.fdr}
\title{perm.fdr}
\usage{
perm.fdr(input_df, perm_df, Pvals_col_name, name)
}
\arguments{
\item{input_df}{no-permutation, anything}

\item{perm_df}{permutation p, only contain p of interesting covariate}

\item{Pvals_col_name}{in input_df, only one}

\item{name}{just for final output, could be same with Pvals_col_name}
}
\value{

}
\description{
Function for calculating permutation-based false discovery rates based on a generalization of the false discovery rate method of Storey and Tibshirani (2003)
}
