% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visual-related.R
\name{plot_gg}
\alias{plot_gg}
\title{plot_gg}
\usage{
plot_gg(gg, row, col)
}
\arguments{
\item{gg}{a vector containing all the ggplot object}

\item{row}{number of row}

\item{col}{number of column}
}
\value{

}
\description{
plot multi-panel ggplot
}
