% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visual-related.R
\name{plot_gsea_comp}
\alias{plot_gsea_comp}
\title{plot_gsea_comp}
\usage{
plot_gsea_comp(data, var1, var2, pcut = 0.05, lab_size = 2.5, ptsize = 0.8)
}
\arguments{
\item{data}{GSEA result}

\item{var1}{in form ES_p_var1, ES_var1}

\item{var2}{in form ES_p_var2, ES_var2}

\item{pcut}{cut-off to label in color}

\item{lab_size}{default 2.5, decrease if too crowded}

\item{ptsize}{point size}
}
\value{
a gg object
}
\description{
compare the ES score for two variables
}
