% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visual-related.R
\name{plot_gsea_comp_twogsea}
\alias{plot_gsea_comp_twogsea}
\title{plot_gsea_comp_twogsea}
\usage{
plot_gsea_comp_twogsea(
  gsealist,
  varvec,
  suffix = c("X", "Y"),
  pcut = 0.05,
  lab_size = 2.5,
  ptsize = 0.8,
  max_overlap = 20
)
}
\arguments{
\item{gsealist}{a list, one or two gsea data frame}

\item{varvec}{a vector, one or two varname}

\item{suffix}{a vector, suffix for var1 and var2}

\item{pcut}{default to 0.05}

\item{lab_size}{label size, default to 2.5}

\item{ptsize}{point size, default to 0.8}

\item{max_overlap}{default to 20}
}
\value{

}
\description{
plot_gsea_comp_twogsea
}
