% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visual-related.R
\name{plot_p}
\alias{plot_p}
\title{plot_p}
\usage{
plot_p(
  dat,
  p_name,
  x_name,
  bw = 0.005,
  plim = 0.05,
  row,
  col,
  x_axis_size = 50,
  savedir = "NA",
  unitwidth = 1440,
  unitheight = 600
)
}
\arguments{
\item{dat}{data frame, emmreml result}

\item{p_name}{a vector containing the name for the p value}

\item{x_name}{a vector containing the label for x axis}

\item{bw}{binwidth, default as 0.005}

\item{plim}{p value color upper limit: default 0.05}

\item{row}{row number}

\item{col}{column number}

\item{x_axis_size}{letter size, default to 50}

\item{savedir}{output png dir, default to "NA"}

\item{unitwidth}{width of a single p histogram, default to 1440}

\item{unitheight}{height of a single p histogram, default to 600}
}
\value{

}
\description{
plot and save p histogram for Emmreml
}
