% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emmreml-related.R
\name{run_emm}
\alias{run_emm}
\title{run_emm}
\usage{
run_emm(z, resids, rel, design, numcore = 3)
}
\arguments{
\item{z}{completed cases}

\item{resids}{GE residual matrix}

\item{rel}{relatedness matrix}

\item{design}{design matrix}

\item{numcore}{number of core to use, default to 3}
}
\value{
a data frame containing emmreml results
}
\description{
run_emm
}
