% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visual-related.R
\name{save_gg_group}
\alias{save_gg_group}
\title{save_gg_group}
\usage{
save_gg_group(gg, row, col, savedir, unitwidth = 1440, unitheight = 600)
}
\arguments{
\item{gg}{a list of gg objects}

\item{row}{nrow}

\item{col}{ncol}

\item{savedir}{outputdir}

\item{unitwidth}{default 1440}

\item{unitheight}{default 600}
}
\value{
nothing, save plot
}
\description{
save_gg_group
}
