% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visual-related.R
\name{stdbeta_plot}
\alias{stdbeta_plot}
\title{stdbeta_plot}
\usage{
stdbeta_plot(x, y, condition = "lps", data, p = 0.1)
}
\arguments{
\item{x}{parameter 1}

\item{y}{parameter 2}

\item{condition}{usually lps or null}

\item{data}{emmreml result, colname in form beta_x_condition, var_x_condition}

\item{p}{p cut-off to mark red}
}
\value{
a ggplot object
}
\description{
std_beta calculaiton and comparison based on emmreml result.
}
