% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visual-related.R
\name{stdbeta_plot_twoemm}
\alias{stdbeta_plot_twoemm}
\title{Plot same effect from two emm result}
\usage{
stdbeta_plot_twoemm(
  emm1,
  emm2,
  var_name,
  suffix = c("X", "Y"),
  pcut = 0.05,
  ptsize = 0.5,
  color_lab = "specificity"
)
}
\arguments{
\item{emm1}{emm1}

\item{emm2}{emm2}

\item{var_name}{var_var_name}

\item{suffix}{var_var_name_suffix}

\item{pcut}{default 0.05}

\item{ptsize}{default 0.5}

\item{color_lab}{default specificity}
}
\value{

}
\description{
Plot same effect from two emm result
}
